[chore #163204269] Split validators into modules
